move submitTransaction to Contract class
